python-backend/requirements.txt
# Polymarket API and WebSocket
websocket-client==1.7.0
requests==2.31.0

# AI Models
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Data Processing
pandas==2.2.0
python-dotenv==1.0.1

# Database
mysql-connector-python==8.3.0

# Async and WebSocket Server
websockets==12.0
aiohttp==3.9.3
asyncio==3.4.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
termcolor==2.4.0
